Add update packages and file structure.
