+++ package.json
@@ -21,16 +21,21 @@
     "prepare": "husky"
   },
   "dependencies": {
+    "@clerk/nextjs": "^5.1.3",
     "@radix-ui/react-popover": "^1.0.7",
+    "@uploadthing/react": "^6.6.0",
     "classnames": "^2.5.1",
     "date-fns": "^3.6.0",
+    "mongoose": "^8.4.1",
     "netlify": "^13.1.16",
     "next": "14.2.3",
     "next-nprogress-bar": "^2.3.12",
     "react": "^18",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18",
-    "swr": "^2.2.5"
+    "svix": "^1.24.0",
+    "swr": "^2.2.5",
+    "uploadthing": "^6.12.0"
   },
   "devDependencies": {
     "@types/node": "^20",
